progress on randomizer

+++ src/solver/IRandomizer.h
@@ -38,7 +38,7 @@ public:
 	virtual ~IRandomizer() { }
 
 	virtual bool randomize(
-			uint64_t								seed,
+			uint32_t								seed,
 			const std::vector<FieldSP>				&fields,
 			const std::vector<ConstraintBlockSP>	&constraints) = 0;
